+++ test/docusign_rest/client_test.rb
@@ -282,6 +282,16 @@ describe DocusignRest::Client do
             .must_equal "SUCCESS"
         end
       end
+      it "should get envelope's sender view" do
+        VCR.use_cassette("get_envelope_sender_view") do
+          response = @client.get_sender_view(
+            envelope_id: @envelope_response["envelopeId"],
+          )
+
+          response['errorCode'].must_be_nil
+          response['url'].must_match(/http/)
+        end
+      end
     end
   end
 end